pyproject.toml
[tool.poetry]
name = "notorious"
version = "0.1.0"
description = ""
authors = ["githubjacky"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
loguru = "<0.7.0"
orjson = "^3.9.4"
tqdm = "^4.66.1"
pandas = "==1.*"
calibrate-google-trends = {git = "https://github.com/githubjacky/calibrate-google-trends.git"}
urllib3 = "==1.*"
networkx = ">=2.4"
openpyxl = "^3.1.2"
pytrends = ">=4.7.2"
ballyregan = "^1.0.6"
fake-useragent = "^1.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
jupyter-bokeh = "^3.0.7"
hvplot = "^0.9.1"
pytest = "^7.4.4"
jupyterlab-execute-time = "2.3.1"
jupyterlab = "3.6.6"
jupyterlab-lsp = "4.3.0"
python-lsp-server = {extras = ["all"], version = "^1.8.2"}
poetry-kernel = "^0.1.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
